+++ src/backend/joanie/tests/core/test_helpers.py
@@ -136,7 +136,7 @@ class HelpersTestCase(TestCase):
         self.assertEqual(certificate_qs.count(), 0)
 
         # DB queries should be minimized
-        with self.assertNumQueries(13):
+        with self.assertNumQueries(14):
             _certificate, created = order.get_or_generate_certificate()
         self.assertTrue(created)
         self.assertEqual(certificate_qs.count(), 1)